[Fix]Add dimensions to MLU copy
To have the same shape of MLU copy and Scatter.forward, a 1-dim list has to be added.

Origin MMDP / MMDDP with this scatter_gather returns a list of tensor for each targets, and here MLU copy returns the same shape Tensor in a list of just one default target.